+++ packages/framework/src/Model/Advert/Advert.php
@@ -13,6 +13,9 @@ class Advert
     const TYPE_IMAGE = 'image';
     const TYPE_CODE = 'code';
 
+    /**
+     * @deprecated Use of position constants is discouraged, use literal strings instead.
+     */
     const POSITION_HEADER = 'header';
     const POSITION_FOOTER = 'footer';
     const POSITION_PRODUCT_LIST = 'productList';